+++ CHANGELOG.md
@@ -13,6 +13,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [v1.12.0](https://raw.githubusercontent.com/Synogun/FarmRPGPlus/refs/tags/v1.12.0/dist/FarmRPGPlus.user.js)
+
+### Added
+- `HomePage` - Adds Track Trading in Borgen's Camp config, which tracks if the player has traded in Borgen's Camp on Wednesdays.
+- `FishingPage` - Adds Fish On Center feature, which centers the fishing shadow on the screen.
+
+### Changed
+
+### Fixed
+
 ## [v1.10.4](https://raw.githubusercontent.com/Synogun/FarmRPGPlus/refs/tags/v1.10.4/dist/FarmRPGPlus.user.js)
 
 ### Added